requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# GitHub models support
openai>=1.0.0
anthropic>=0.40.0

# Search and data
tavily-python>=0.5.0
# duckduckgo-search>=6.0.0  # Alternativa gratuita, descomentar si se necesita
beautifulsoup4>=4.12.0
requests>=2.32.0

# Database and storage
chromadb>=0.4.0  # Versión más estable
sqlalchemy>=2.0.0
# sqlite-vss>=0.1.0  # Opcional, puede dar problemas en Windows

# Data processing
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
# asyncio viene incluido en Python 3.7+

# Visualization (optional but useful)
matplotlib>=3.8.0
plotly>=5.20.0
streamlit>=1.32.0

# Utils
rich>=13.0.0  # For pretty console output
tqdm>=4.66.0  # Progress bars
loguru>=0.7.0  # Better logging